requirements.txt
python-nmap
# TODO - first release
git+https://github.com/OpenVoiceOS/ovos-PHAL-plugin-commonIOT
